requirements.txt
# Core dependencies
mammoth>=1.6.0
markdownify>=0.11.0
pandas>=2.0.0
pdfminer.six>=20221105
pdfplumber>=0.10.0
python-pptx>=0.6.21
puremagic>=1.15
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.3

# PDF Processing
camelot-py>=0.10.1
tabula-py>=2.7.0
pdf2image>=1.16.3
opencv-python>=4.8.0
ghostscript>=0.7
JPype1>=1.4.1
PyMuPDF>=1.21.0

# Image Processing
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0

# UI
PyQt6>=6.4.0
keyring>=24.0.0

# Progress Bars
tqdm>=4.65.0

# AI/ML Support
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
openai>=1.3.0 